Use munit in LastSuite
